Split search terms only for __contains queries
Split search terms only for `__contains` queries and not for
`__startswith` or `__endswith`. We no also split not only be
whitespace but tab and newline.

If multiple search fields are defined conditions combined with
and OR. If single word matches in a for contains queries are
OR combined as well.

Co-Authored-By: codingjoe <info@johanneshoppe.com>